Add rudimentary support for Catalina screen recording permissions
macOS Catalina requires a new permission to read the window title from
window enumeration APIs. The symptom of the problem of running without
permissions is that the kCGWindowName field in the dictionaries returned
by CGWindowListCopyWindowInfo() is not populated.

There does not appear to be a clean check for this permission short of
actually trying to record the screen.

For now, just blindly try to record a pixel from the screen to trigger
the permission dialog, as described here:
https://stackoverflow.com/questions/59337022/enabling-screen-recording-api-in-catalina-kcgwindowname

There are TODOs in the code to actually check for whether recording is
enabled, and to return true in that case.