Update screen recording permission check to actually try recording
Calling CGDisplayStreamCreate() to actually try to record from the screen
appears to be a safe way to check if screen recording is enabled, and to
trigger a user consent dialog the first time the program is run. Update
isAuthorizedForScreenRecording() to check this way instead of trying to grab
a pixel from the desktop (which triggers the consent dialog the first time the
program is run, but always returns success).

Also, make the check conditional on OS X version.